+++ shopify/packages/ajax-cart-components/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brytdesigns/shopify-ajax-cart-components
 
+## 1.1.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/shopify-ajax-cart@1.0.20
+
 ## 1.1.18
 
 ### Patch Changes